model/encoder_decoder/requirement.txt
keras == 2.13.1
keras-tuner == 1.4.6
numpy == 1.24.3
pandas == 1.5.2
tenacity == 8.0.1
tensorflow == 2.13.0
google-cloud-storage == 2.15.0
scikit-learn == 1.4.1.post1